some minor reorganization
