+++ test_vault/.obsidian/community-plugins.json
@@ -1,7 +1,7 @@
 [
   "dataview",
-  "supercharged-links-obsidian",
   "obsidian-style-settings",
   "mathlinks",
-  "obsidian-math-booster"
+  "obsidian-math-booster",
+  "supercharged-links-obsidian"
 ]
